requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-crispy-forms==2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
